requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pandas>=2.2.2
matplotlib>=3.8.4
tenacity>=8.2.3

